requirements.txt
python-telegram-bot==21.4
fastapi==0.115.2
uvicorn==0.30.6
pydantic==2.9.2
httpx==0.27.2
Pillow==10.4.0
